@@ -1,21 +1,20 @@
 #!/usr/bin/env node
 
+import process from 'node:process';
 import meow from 'meow';
 import chalk from 'chalk';
-import {Flags, getConfig} from './config.js';
+import {type Flags, getConfig} from './config.js';
 import {Format, Logger, LogLevel} from './logger.js';
 import {
-  LinkChecker,
-  LinkState,
-  LinkResult,
-  CheckOptions,
-  RetryInfo,
+	LinkChecker,
+	LinkState,
+	type LinkResult,
+	type CheckOptions,
+	type RetryInfo,
 } from './index.js';
 
-/* eslint-disable no-process-exit */
-
 const cli = meow(
-  `
+	`
     Usage
       $ linkinator LOCATION [ --arguments ]
 
@@ -88,273 +87,322 @@ const cli = meow(
       $ linkinator . --skip www.googleapis.com
       $ linkinator . --format CSV
 `,
-  {
-    importMeta: import.meta,
-    flags: {
-      config: {type: 'string'},
-      concurrency: {type: 'number'},
-      recurse: {type: 'boolean', shortFlag: 'r'},
-      skip: {type: 'string', shortFlag: 's', isMultiple: true},
-      format: {type: 'string', shortFlag: 'f'},
-      silent: {type: 'boolean'},
-      timeout: {type: 'number'},
-      markdown: {type: 'boolean'},
-      serverRoot: {type: 'string'},
-      verbosity: {type: 'string'},
-      directoryListing: {type: 'boolean'},
-      retry: {type: 'boolean'},
-      retryErrors: {type: 'boolean'},
-      retryErrorsCount: {type: 'number', default: 5},
-      retryErrorsJitter: {type: 'number', default: 3000},
-      urlRewriteSearch: {type: 'string'},
-      urlReWriteReplace: {type: 'string'},
-    },
-    booleanDefault: undefined,
-  }
+	{
+		importMeta: import.meta,
+		flags: {
+			config: {type: 'string'},
+			concurrency: {type: 'number'},
+			recurse: {type: 'boolean', shortFlag: 'r'},
+			skip: {type: 'string', shortFlag: 's', isMultiple: true},
+			format: {type: 'string', shortFlag: 'f'},
+			silent: {type: 'boolean'},
+			timeout: {type: 'number'},
+			markdown: {type: 'boolean'},
+			serverRoot: {type: 'string'},
+			verbosity: {type: 'string'},
+			directoryListing: {type: 'boolean'},
+			retry: {type: 'boolean'},
+			retryErrors: {type: 'boolean'},
+			retryErrorsCount: {type: 'number', default: 5},
+			retryErrorsJitter: {type: 'number', default: 3000},
+			urlRewriteSearch: {type: 'string'},
+			urlReWriteReplace: {type: 'string'},
+		},
+		booleanDefault: undefined,
+	},
 );
 
 let flags: Flags;
 
 async function main() {
-  if (cli.input.length < 1) {
-    cli.showHelp();
-    return;
-  }
-  flags = await getConfig(cli.flags);
-  if (
-    (flags.urlRewriteReplace && !flags.urlRewriteSearch) ||
-    (flags.urlRewriteSearch && !flags.urlRewriteReplace)
-  ) {
-    throw new Error(
-      'The url-rewrite-replace flag must be used with the url-rewrite-search flag.'
-    );
-  }
+	if (cli.input.length === 0) {
+		cli.showHelp();
+		return;
+	}
 
-  const start = Date.now();
-  const verbosity = parseVerbosity(flags);
-  const format = parseFormat(flags);
-  const logger = new Logger(verbosity, format);
+	flags = await getConfig(cli.flags);
+	if (
+		(flags.urlRewriteReplace && !flags.urlRewriteSearch) ||
+		(flags.urlRewriteSearch && !flags.urlRewriteReplace)
+	) {
+		throw new Error(
+			'The url-rewrite-replace flag must be used with the url-rewrite-search flag.',
+		);
+	}
 
-  logger.error(`🏊‍♂️ crawling ${cli.input}`);
+	const start = Date.now();
+	const verbosity = parseVerbosity(flags);
+	const format = parseFormat(flags);
+	const logger = new Logger(verbosity, format);
 
-  const checker = new LinkChecker();
-  if (format === Format.CSV) {
-    const header = 'url,status,state,parent,failureDetails';
-    console.log(header);
-  }
-  checker.on('retry', (info: RetryInfo) => {
-    logger.warn(`Retrying: ${info.url} in ${info.secondsUntilRetry} seconds.`);
-  });
-  checker.on('link', (link: LinkResult) => {
-    let state = '';
-    switch (link.state) {
-      case LinkState.BROKEN:
-        state = `[${chalk.red(link.status!.toString())}]`;
-        logger.error(`${state} ${chalk.gray(link.url)}`);
-        break;
-      case LinkState.OK:
-        state = `[${chalk.green(link.status!.toString())}]`;
-        logger.warn(`${state} ${chalk.gray(link.url)}`);
-        break;
-      case LinkState.SKIPPED:
-        state = `[${chalk.grey('SKP')}]`;
-        logger.info(`${state} ${chalk.gray(link.url)}`);
-        break;
-    }
-    if (format === Format.CSV) {
-      const showIt = shouldShowResult(link, verbosity);
-      if (showIt) {
-        console.log(
-          `"${link.url}",${link.status},${link.state},"${link.parent || ''}","${
-            link.failureDetails || ''
-          }"`
-        );
-      }
-    }
-  });
-  const opts: CheckOptions = {
-    path: cli.input,
-    recurse: flags.recurse,
-    timeout: Number(flags.timeout),
-    markdown: flags.markdown,
-    concurrency: Number(flags.concurrency),
-    serverRoot: flags.serverRoot,
-    directoryListing: flags.directoryListing,
-    retry: flags.retry,
-    retryErrors: flags.retryErrors,
-    retryErrorsCount: Number(flags.retryErrorsCount),
-    retryErrorsJitter: Number(flags.retryErrorsJitter),
-  };
-  if (flags.skip) {
-    if (typeof flags.skip === 'string') {
-      opts.linksToSkip = flags.skip.split(/[\s,]+/).filter(x => !!x);
-    } else if (Array.isArray(flags.skip)) {
-      // with `isMultiple` enabled in meow, a comma delimeted list will still
-      // be passed as an array, but with a single element that still needs to
-      // be split.
-      opts.linksToSkip = [];
-      for (const skip of flags.skip) {
-        const rules = skip.split(/[\s,]+/).filter(x => !!x);
-        opts.linksToSkip.push(...rules);
-      }
-    }
-  }
-  if (flags.urlRewriteSearch && flags.urlRewriteReplace) {
-    opts.urlRewriteExpressions = [
-      {
-        pattern: new RegExp(flags.urlRewriteSearch),
-        replacement: flags.urlRewriteReplace,
-      },
-    ];
-  }
-  const result = await checker.check(opts);
-  const filteredResults = result.links.filter(link =>
-    shouldShowResult(link, verbosity)
-  );
-  if (format === Format.JSON) {
-    result.links = filteredResults;
-    console.log(JSON.stringify(result, null, 2));
-    return;
-  } else if (format === Format.CSV) {
-    return;
-  } else {
-    // Build a collection scanned links, collated by the parent link used in
-    // the scan.  For example:
-    // {
-    //   "./README.md": [
-    //     {
-    //       url: "https://img.shields.io/npm/v/linkinator.svg",
-    //       status: 200
-    //       ....
-    //     }
-    //   ],
-    // }
-    const parents = result.links.reduce((acc, curr) => {
-      const parent = curr.parent || '';
-      if (!acc[parent]) {
-        acc[parent] = [];
-      }
-      acc[parent].push(curr);
-      return acc;
-    }, {} as {[index: string]: LinkResult[]});
+	logger.error(`🏊‍♂️ crawling ${cli.input.join(' ')}`);
 
-    Object.keys(parents).forEach(parent => {
-      // prune links based on verbosity
-      const links = parents[parent].filter(link => {
-        if (verbosity === LogLevel.NONE) {
-          return false;
-        }
-        if (link.state === LinkState.BROKEN) {
-          return true;
-        }
-        if (link.state === LinkState.OK) {
-          if (verbosity <= LogLevel.WARNING) {
-            return true;
-          }
-        }
-        if (link.state === LinkState.SKIPPED) {
-          if (verbosity <= LogLevel.INFO) {
-            return true;
-          }
-        }
-        return false;
-      });
-      if (links.length === 0) {
-        return;
-      }
-      logger.error(chalk.blue(parent));
-      links.forEach(link => {
-        let state = '';
-        switch (link.state) {
-          case LinkState.BROKEN:
-            state = `[${chalk.red(link.status!.toString())}]`;
-            logger.error(`  ${state} ${chalk.gray(link.url)}`);
-            logger.debug(JSON.stringify(link.failureDetails, null, 2));
-            break;
-          case LinkState.OK:
-            state = `[${chalk.green(link.status!.toString())}]`;
-            logger.warn(`  ${state} ${chalk.gray(link.url)}`);
-            break;
-          case LinkState.SKIPPED:
-            state = `[${chalk.grey('SKP')}]`;
-            logger.info(`  ${state} ${chalk.gray(link.url)}`);
-            break;
-        }
-      });
-    });
-  }
+	const checker = new LinkChecker();
+	if (format === Format.CSV) {
+		const header = 'url,status,state,parent,failureDetails';
+		console.log(header);
+	}
 
-  const total = (Date.now() - start) / 1000;
-  const scannedLinks = result.links.filter(x => x.state !== LinkState.SKIPPED);
-  if (!result.passed) {
-    const borked = result.links.filter(x => x.state === LinkState.BROKEN);
-    logger.error(
-      chalk.bold(
-        `${chalk.red('ERROR')}: Detected ${
-          borked.length
-        } broken links. Scanned ${chalk.yellow(
-          scannedLinks.length.toString()
-        )} links in ${chalk.cyan(total.toString())} seconds.`
-      )
-    );
-    process.exit(1);
-  }
+	checker.on('retry', (info: RetryInfo) => {
+		logger.warn(`Retrying: ${info.url} in ${info.secondsUntilRetry} seconds.`);
+	});
+	checker.on('link', (link: LinkResult) => {
+		let state = '';
+		switch (link.state) {
+			case LinkState.BROKEN: {
+				state = `[${chalk.red(link.status!.toString())}]`;
+				logger.error(`${state} ${chalk.gray(link.url)}`);
+				break;
+			}
 
-  logger.error(
-    chalk.bold(
-      `🤖 Successfully scanned ${chalk.green(
-        scannedLinks.length.toString()
-      )} links in ${chalk.cyan(total.toString())} seconds.`
-    )
-  );
+			case LinkState.OK: {
+				state = `[${chalk.green(link.status!.toString())}]`;
+				logger.warn(`${state} ${chalk.gray(link.url)}`);
+				break;
+			}
+
+			case LinkState.SKIPPED: {
+				state = `[${chalk.grey('SKP')}]`;
+				logger.info(`${state} ${chalk.gray(link.url)}`);
+				break;
+			}
+
+			default: {
+				throw new Error('Unexpected link state.');
+			}
+		}
+
+		if (format === Format.CSV) {
+			const showIt = shouldShowResult(link, verbosity);
+			if (showIt) {
+				console.log(
+					`"${link.url}",${link.status},${link.state},"${link.parent || ''}","${
+						link.failureDetails?.toString() || ''
+					}"`,
+				);
+			}
+		}
+	});
+	const options: CheckOptions = {
+		path: cli.input,
+		recurse: flags.recurse,
+		timeout: Number(flags.timeout),
+		markdown: flags.markdown,
+		concurrency: Number(flags.concurrency),
+		serverRoot: flags.serverRoot,
+		directoryListing: flags.directoryListing,
+		retry: flags.retry,
+		retryErrors: flags.retryErrors,
+		retryErrorsCount: Number(flags.retryErrorsCount),
+		retryErrorsJitter: Number(flags.retryErrorsJitter),
+	};
+	if (flags.skip) {
+		if (typeof flags.skip === 'string') {
+			options.linksToSkip = flags.skip.split(/[\s,]+/).filter(Boolean);
+		} else if (Array.isArray(flags.skip)) {
+			// With `isMultiple` enabled in meow, a comma delimeted list will still
+			// be passed as an array, but with a single element that still needs to
+			// be split.
+			options.linksToSkip = [];
+			for (const skip of flags.skip) {
+				const rules = skip.split(/[\s,]+/).filter(Boolean);
+				options.linksToSkip.push(...rules);
+			}
+		}
+	}
+
+	if (flags.urlRewriteSearch && flags.urlRewriteReplace) {
+		options.urlRewriteExpressions = [
+			{
+				pattern: new RegExp(flags.urlRewriteSearch),
+				replacement: flags.urlRewriteReplace,
+			},
+		];
+	}
+
+	const result = await checker.check(options);
+	const filteredResults = result.links.filter((link) =>
+		shouldShowResult(link, verbosity),
+	);
+	if (format === Format.JSON) {
+		result.links = filteredResults;
+		console.log(JSON.stringify(result, null, 2));
+		return;
+	}
+
+	if (format === Format.CSV) {
+		return;
+	}
+
+	// Build a collection scanned links, collated by the parent link used in
+	// the scan.  For example:
+	// {
+	//   "./README.md": [
+	//     {
+	//       url: "https://img.shields.io/npm/v/linkinator.svg",
+	//       status: 200
+	//       ....
+	//     }
+	//   ],
+	// }
+	// eslint-disable-next-line unicorn/no-array-reduce
+	const parents = result.links.reduce<Record<string, LinkResult[]>>(
+		(acc, curr) => {
+			const parent = curr.parent || '';
+			if (!acc[parent]) {
+				acc[parent] = [];
+			}
+
+			acc[parent].push(curr);
+			return acc;
+		},
+		{},
+	);
+
+	for (const parent of Object.keys(parents)) {
+		// Prune links based on verbosity
+		const links = parents[parent].filter((link) => {
+			if (verbosity === LogLevel.NONE) {
+				return false;
+			}
+
+			if (link.state === LinkState.BROKEN) {
+				return true;
+			}
+
+			if (link.state === LinkState.OK && verbosity <= LogLevel.WARNING) {
+				return true;
+			}
+
+			if (link.state === LinkState.SKIPPED && verbosity <= LogLevel.INFO) {
+				return true;
+			}
+
+			return false;
+		});
+		if (links.length === 0) {
+			continue;
+		}
+
+		logger.error(chalk.blue(parent));
+		for (const link of links) {
+			let state = '';
+			switch (link.state) {
+				case LinkState.BROKEN: {
+					state = `[${chalk.red(link.status!.toString())}]`;
+					logger.error(`  ${state} ${chalk.gray(link.url)}`);
+					logger.debug(JSON.stringify(link.failureDetails, null, 2));
+					break;
+				}
+
+				case LinkState.OK: {
+					state = `[${chalk.green(link.status!.toString())}]`;
+					logger.warn(`  ${state} ${chalk.gray(link.url)}`);
+					break;
+				}
+
+				case LinkState.SKIPPED: {
+					state = `[${chalk.grey('SKP')}]`;
+					logger.info(`  ${state} ${chalk.gray(link.url)}`);
+					break;
+				}
+
+				default: {
+					throw new Error('Unexpected link state.');
+				}
+			}
+		}
+	}
+
+	const total = (Date.now() - start) / 1000;
+	const scannedLinks = result.links.filter(
+		(x) => x.state !== LinkState.SKIPPED,
+	);
+	if (!result.passed) {
+		const borked = result.links.filter((x) => x.state === LinkState.BROKEN);
+		logger.error(
+			chalk.bold(
+				`${chalk.red('ERROR')}: Detected ${
+					borked.length
+				} broken links. Scanned ${chalk.yellow(
+					scannedLinks.length.toString(),
+				)} links in ${chalk.cyan(total.toString())} seconds.`,
+			),
+		);
+		process.exit(1);
+	}
+
+	logger.error(
+		chalk.bold(
+			`🤖 Successfully scanned ${chalk.green(
+				scannedLinks.length.toString(),
+			)} links in ${chalk.cyan(total.toString())} seconds.`,
+		),
+	);
 }
 
 function parseVerbosity(flags: Flags): LogLevel {
-  if (flags.silent && flags.verbosity) {
-    throw new Error(
-      'The SILENT and VERBOSITY flags cannot both be defined. Please consider using VERBOSITY only.'
-    );
-  }
-  if (flags.silent) {
-    return LogLevel.ERROR;
-  }
-  if (!flags.verbosity) {
-    return LogLevel.WARNING;
-  }
-  const verbosity = flags.verbosity.toUpperCase();
-  const options = Object.values(LogLevel);
-  if (!options.includes(verbosity)) {
-    throw new Error(
-      `Invalid flag: VERBOSITY must be one of [${options.join(',')}]`
-    );
-  }
-  return LogLevel[verbosity as keyof typeof LogLevel];
+	if (flags.silent && flags.verbosity) {
+		throw new Error(
+			'The SILENT and VERBOSITY flags cannot both be defined. Please consider using VERBOSITY only.',
+		);
+	}
+
+	if (flags.silent) {
+		return LogLevel.ERROR;
+	}
+
+	if (!flags.verbosity) {
+		return LogLevel.WARNING;
+	}
+
+	const verbosity = flags.verbosity.toUpperCase();
+	const options = Object.values(LogLevel);
+	if (!options.includes(verbosity)) {
+		throw new Error(
+			`Invalid flag: VERBOSITY must be one of [${options.join(',')}]`,
+		);
+	}
+
+	return LogLevel[verbosity as keyof typeof LogLevel];
 }
 
 function parseFormat(flags: Flags): Format {
-  if (!flags.format) {
-    return Format.TEXT;
-  }
-  flags.format = flags.format.toUpperCase();
-  const options = Object.values(Format);
-  if (!options.includes(flags.format)) {
-    throw new Error("Invalid flag: FORMAT must be 'TEXT', 'JSON', or 'CSV'.");
-  }
-  return Format[flags.format as keyof typeof Format];
+	if (!flags.format) {
+		return Format.TEXT;
+	}
+
+	flags.format = flags.format.toUpperCase();
+	const options = Object.values(Format);
+	if (!options.includes(flags.format)) {
+		throw new Error("Invalid flag: FORMAT must be 'TEXT', 'JSON', or 'CSV'.");
+	}
+
+	return Format[flags.format as keyof typeof Format];
 }
 
 function shouldShowResult(link: LinkResult, verbosity: LogLevel) {
-  switch (link.state) {
-    case LinkState.OK:
-      return verbosity <= LogLevel.WARNING;
-    case LinkState.BROKEN:
-      if (verbosity > LogLevel.DEBUG) {
-        link.failureDetails = undefined;
-      }
-      return verbosity <= LogLevel.ERROR;
-    case LinkState.SKIPPED:
-      return verbosity <= LogLevel.INFO;
-  }
+	switch (link.state) {
+		case LinkState.OK: {
+			return verbosity <= LogLevel.WARNING;
+		}
+
+		case LinkState.BROKEN: {
+			if (verbosity > LogLevel.DEBUG) {
+				link.failureDetails = undefined;
+			}
+
+			return verbosity <= LogLevel.ERROR;
+		}
+
+		case LinkState.SKIPPED: {
+			return verbosity <= LogLevel.INFO;
+		}
+
+		default: {
+			throw new Error('Unexpected link state.');
+		}
+	}
 }
 
-main();
+await main();
