@@ -1,8 +1,8 @@
 export default {
-  format: 'json',
-  recurse: true,
-  silent: true,
-  concurrency: 17,
-  skip: '🪐',
-  directoryListing: false,
+	format: 'json',
+	recurse: true,
+	silent: true,
+	concurrency: 17,
+	skip: '🪐',
+	directoryListing: false,
 };
